moved everything over to the issuelabelerservice
