+++ tools/issue-labeler/IssueLabeler.sln
@@ -9,9 +9,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IssueLabelerService", "src\IssueLabelerService\IssueLabelerService.csproj", "{4C9E75AF-468F-4DF7-BACD-EC0C2C66A96F}"
-	ProjectSection(ProjectDependencies) = postProject
-		{38F6D79F-3BAB-755C-9532-78FE4D968AE4} = {38F6D79F-3BAB-755C-9532-78FE4D968AE4}
-	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SearchIndexCreator", "src\SearchIndexCreator\SearchIndexCreator.csproj", "{0AEAF8DD-C370-4090-B439-9CF364D29869}"
 	ProjectSection(ProjectDependencies) = postProject
@@ -29,12 +26,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IssueLabeler.Shared", "src\IssueLabeler.Shared\IssueLabeler.Shared.csproj", "{9E6BA2D8-3BBE-40D6-9DAF-0FC0CD362BD4}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hubbup.MikLabelModel.Tests", "tests\Hubbup.MikLabelModel.Tests\Hubbup.MikLabelModel.Tests.csproj", "{CD3F13F1-8890-490A-BB47-9382E2131F5D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AzureRagService", "src\AzureRagService\AzureRagService.csproj", "{607AE5A6-7B8B-6E10-577E-D2692BEBAA70}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LabelerFactory", "src\LabelerFactory\LabelerFactory.csproj", "{950DADD7-CF21-0717-ECB2-1347E0E3E182}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConfigurationService", "src\ConfigurationService\ConfigurationService.csproj", "{38F6D79F-3BAB-755C-9532-78FE4D968AE4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -74,18 +65,6 @@ Global
 		{CD3F13F1-8890-490A-BB47-9382E2131F5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CD3F13F1-8890-490A-BB47-9382E2131F5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CD3F13F1-8890-490A-BB47-9382E2131F5D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{607AE5A6-7B8B-6E10-577E-D2692BEBAA70}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{607AE5A6-7B8B-6E10-577E-D2692BEBAA70}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{607AE5A6-7B8B-6E10-577E-D2692BEBAA70}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{607AE5A6-7B8B-6E10-577E-D2692BEBAA70}.Release|Any CPU.Build.0 = Release|Any CPU
-		{950DADD7-CF21-0717-ECB2-1347E0E3E182}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{950DADD7-CF21-0717-ECB2-1347E0E3E182}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{950DADD7-CF21-0717-ECB2-1347E0E3E182}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{950DADD7-CF21-0717-ECB2-1347E0E3E182}.Release|Any CPU.Build.0 = Release|Any CPU
-		{38F6D79F-3BAB-755C-9532-78FE4D968AE4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{38F6D79F-3BAB-755C-9532-78FE4D968AE4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{38F6D79F-3BAB-755C-9532-78FE4D968AE4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{38F6D79F-3BAB-755C-9532-78FE4D968AE4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
