+++ tools/issue-labeler/src/IssueLabelerService/IssueLabelerService.csproj
@@ -21,11 +21,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AzureRagService\AzureRagService.csproj" />
     <ProjectReference Include="..\Hubbup.MikLabelModel\Hubbup.MikLabelModel.csproj" />
     <ProjectReference Include="..\IssueLabeler.Shared\IssueLabeler.Shared.csproj" />
-    <ProjectReference Include="..\ConfigurationService\ConfigurationService.csproj" />
-    <ProjectReference Include="..\LabelerFactory\LabelerFactory.csproj"></ProjectReference>
   </ItemGroup>
 
   <ItemGroup>
