+++ tools/issue-labeler/src/IssueLabelerService/Program.cs
@@ -4,7 +4,6 @@
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using AzureRagService;
 using Hubbup.MikLabelModel;
 using Azure.Identity;
 using System;
@@ -12,9 +11,8 @@ using Azure.AI.OpenAI;
 using OpenAI.Chat;
 using Azure.Search.Documents.Indexes;
 using Microsoft.Extensions.Configuration;
-using ConfigurationService;
 using Azure.Core;
-using LabelerFactory;
+using IssueLabelerService;
 
 var host = new HostBuilder()
     .ConfigureFunctionsWebApplication()
