+++ tools/issue-labeler/src/IssueLabelerService/RagLabeler.cs
@@ -1,10 +1,10 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LabelerFactory
+namespace IssueLabelerService
 {
     internal class RagLabeler
     {
