+++ tools/issue-labeler/src/LabelerFactory/LabelerFactory.csproj
@@ -8,7 +8,6 @@
   <ItemGroup>
     <ProjectReference Include="..\Hubbup.MikLabelModel\Hubbup.MikLabelModel.csproj" />
     <ProjectReference Include="..\IssueLabeler.Shared\IssueLabeler.Shared.csproj" />
-    <ProjectReference Include="..\ConfigurationService\ConfigurationService.csproj" />
   </ItemGroup>
 
 </Project>
