+++ tools/issue-labeler/src/SearchIndexCreator/SearchIndexCreator.csproj
@@ -20,7 +20,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\github-event-processor\Azure.Sdk.Tools.GitHubEventProcessor\Azure.Sdk.Tools.GitHubEventProcessor.csproj" />
-    <ProjectReference Include="..\AzureRagService\AzureRagService.csproj" />
+    <ProjectReference Include="..\IssueLabelerService\IssueLabelerService.csproj" />
   </ItemGroup>
 
   <ItemGroup>
